Added support for read-only databases

+++ sqlite/sqlStorage.py
@@ -41,7 +41,7 @@ class SQLStorage(object):
         db.isolation_level = "DEFERRED"
 
         self.db = db
-        self.cursor = db.cursor()
+        self._cursor = db.cursor()
         self.initialize()
 
         if self.dbid is None:
@@ -53,35 +53,43 @@ class SQLStorage(object):
     #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
     def close(self):
-        self.cursor = None
+        self._cursor = None
         self.db.close()
         self.db = None
 
     def initialize(self):
+        writeOps = 0
+        cursor = self.readCursor
         for attrName in self._sql_init:
             sqlOps = getattr(sqlCreateStorage, attrName)
             for sql in sqlOps:
                 if isinstance(sql, tuple):
                     sqlGuard, sql = sql
 
-                    try:
-                        self.cursor.executescript(sqlGuard)
+                    try: cursor.executescript(sqlGuard)
                     except sqlite3.OperationalError, e: pass
                     else: continue
 
-                try:
-                    self.cursor.executescript(sql)
-                except sqlite3.OperationalError, e:
-                    print >> sys.stderr, "Error while executing creation script: %r on db: %s" % (attrName, self.dbFilename)
-                    raise
+                    try: 
+                        cursor.executescript(sql)
+                        writeOps += 1
+                    except sqlite3.OperationalError, e:
+                        print >> sys.stderr, "Error while executing creation script: %r on db: %s" % (attrName, self.dbFilename)
+                        raise
+                else:
+                    try: cursor.executescript(sql)
+                    except sqlite3.OperationalError, e:
+                        print >> sys.stderr, "Error while executing initialization script: %r on db: %s" % (attrName, self.dbFilename)
+                        raise
 
         self.fetchMetadata()
         self.nextOid = self.getMetaAttr('nextOid', 1000)
-        self.newSession()
-        self.gcInit()
+        if writeOps:
+            self.newSession(cursor)
 
     def fetchMetadata(self):
-        r = self.cursor.execute('select attr, value from odb_metadata')
+        self._metadata = dict()
+        r = self.readCursor.execute('select attr, value from odb_metadata')
         self._metadata = dict(r.fetchall())
 
     def getMetadata(self):
@@ -90,14 +98,19 @@ class SQLStorage(object):
         return self._metadata.get(attr, default)
     def setMetaAttr(self, attr, value):
         r = self._metadata
-        if r.get(attr, object()) != value:
-            r[attr] = value
-            self.cursor.execute(
-                'replace into odb_metadata '
-                '  (attr, value) values (?, ?)', 
-                (attr, value))
+        if r.get(attr, object()) == value:
+            return False
+
+        r[attr] = value
+        r = self.writeCursor
+        r.execute(
+            'replace into odb_metadata '
+            '  (attr, value) values (?, ?)', 
+            (attr, value))
+        return True
     def delMetaAttr(self, attr):
-        r = self.cursor.execute(
+        r = self.writeCursor
+        r.execute(
             'delete from odb_metadata '
             '  where attr=?', (attr,))
         return r.rowcount > 0
@@ -105,19 +118,43 @@ class SQLStorage(object):
     def getDbid(self):
         return self.getMetaAttr('dbid')
     def setDbid(self, dbid):
-        if dbid != self.dbid:
-            return self.setMetaAttr('dbid', dbid)
+        return self.setMetaAttr('dbid', dbid)
     dbid = property(getDbid, setDbid)
 
-    def newSession(self):
-        self.commit()
+    def getWritable(self):
+        return self.getMetaAttr('writable', True)
+    def setWritable(self, writable=True):
+        return self.setMetaAttr('writable', writable)
+    writable = property(getWritable, setWritable)
 
-        self.session = uuid.uuid4() # new random uuid
-        r = self.cursor.execute(
-            'insert into odb_sessions values (NULL, ?, ?)',
-            (str(self.session), self.nextOid))
-        self.ssid = r.lastrowid
-        self.commit()
+    session = None
+    def newSession(self, writeCursor):
+        session = self.session
+        if session is None and self.writable:
+            session = uuid.uuid4() # new random uuid
+            self.session = session
+            print 'writable session:', (session, self.dbFilename)
+            r = writeCursor.execute(
+                'insert into odb_sessions values (NULL, ?, ?)',
+                (str(session), self.nextOid))
+            self.ssid = r.lastrowid
+            self.commit()
+        return session
+
+    #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+    def getReadCursor(self):
+        r = self._cursor
+        return r
+    readCursor = property(getReadCursor)
+
+    def getWriteCursor(self):
+        r = self._cursor
+        if self.session is None:
+            if not self.newSession(r):
+                r = None
+        return r
+    writeCursor = property(getWriteCursor)
 
     #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
@@ -128,17 +165,17 @@ class SQLStorage(object):
     #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
     def allOids(self):
-        r = self.cursor.execute(
+        r = self.readCursor.execute(
             'select oid from oid_lookup')
         return [e[0] for e in r.fetchall()]
 
     def allOidInfo(self):
-        r = self.cursor.execute(
+        r = self.readCursor.execute(
             'select * from oid_lookup')
         return r.fetchall()
 
     def getOidInfo(self, oid):
-        r = self.cursor.execute(
+        r = self.readCursor.execute(
             'select stg_kind, otype '
             '  from oid_lookup where oid=?', (oid,))
         return r.fetchone()
@@ -153,31 +190,33 @@ class SQLStorage(object):
         if not isinstance(stg_kind, str): 
             raise ValueError("Object storage stg_kind must be specified")
 
-        self.cursor.execute(
+        r = self.writeCursor
+        r.execute(
             'replace into oid_lookup (oid, stg_kind, otype, ssid) '
             '  values(?, ?, ?, ?)', (oid, stg_kind, otype, self.ssid))
         return oid
 
     def removeOid(self, oid):
-        self.cursor.execute(
+        r = self.writeCursor
+        r.execute(
             'delete from oid_lookup where oid=?', (oid,))
-        self.cursor.execute(
+        r.execute(
             'delete from exports where oid_ref=?', (oid,))
-        self.cursor.execute(
+        r.execute(
             'delete from mappings where oid_host=?', (oid,))
 
     def allURLPaths(self, incOid=True):
         if incOid:
-            r = self.cursor.execute(
+            r = self.readCursor.execute(
                 "select urlpath, oid_ref from exports")
             return r.fetchall()
         else:
-            r = self.cursor.execute(
+            r = self.readCursor.execute(
                 "select urlpath from exports")
             return [e[0] for e in r.fetchall()]
 
     def getAtURLPath(self, urlpath):
-        r = self.cursor.execute(
+        r = self.readCursor.execute(
             "select oid, stg_kind, otype from exports_lookup"
             "  where urlpath=?", (urlpath,))
         return r.fetchone()
@@ -186,7 +225,7 @@ class SQLStorage(object):
         if not urlpath: 
             return
 
-        r = self.cursor
+        r = self.writeCursor
         r.execute(
             "replace into exports (urlpath, oid_ref, ssid)"
             "  values(?,?,?)", (urlpath, oid, self.ssid))
@@ -194,7 +233,7 @@ class SQLStorage(object):
             """insert into oidGraphMembers values (?)""", (oid,))
 
     def findLiteral(self, value, value_hash, value_type):
-        r = self.cursor.execute(
+        r = self.readCursor.execute(
             'select oid from literals '
             '  where value=? and value_hash=? and value_type=?',
             (value, value_hash, value_type))
@@ -203,13 +242,13 @@ class SQLStorage(object):
             return r[0]
 
     def getLiteralAndType(self, oid):
-        r = self.cursor.execute(
+        r = self.readCursor.execute(
             'select value, value_type '
             '  from literals where oid=?', (oid,))
         return r.fetchone()
 
     def getLiteral(self, oid):
-        r = self.cursor.execute(
+        r = self.readCursor.execute(
             'select value from literals ' 
             '  where oid=?', (oid,))
         r = r.fetchone()
@@ -220,44 +259,47 @@ class SQLStorage(object):
         if oid is not None:
             return oid
 
+        r = self.writeCursor
         oid = self.setOid(None, stg_kind, value_type)
-        self.cursor.execute(
+        r.execute(
             'insert into literals (oid, value, value_type, value_hash, ssid)'
             '  values(?, ?, ?, ?, ?)', (oid, value, value_type, value_hash, self.ssid))
         return oid
 
     def getExternal(self, oid):
-        r = self.cursor.execute(
+        r = self.readCursor.execute(
             'select url from externals '
             '  where oid=?', (oid,))
         r = r.fetchone()
         if r is not None:
             return r[0]
     def setExternal(self, oid, url):
-        self.cursor.execute(
+        r = self.writeCursor
+        r.execute(
             'insert into externals (oid, url, ssid)'
             '  values(?, ?, ?)', (oid, url, self.ssid))
         return oid
 
     def getWeakref(self, oid):
-        r = self.cursor.execute(
+        r = self.readCursor.execute(
             'select v_oid, v_stg_kind, v_otype from weakrefs_lookup '
             '  where oid=?', (oid,))
         return r.fetchone()
     def setWeakref(self, oid, oid_ref):
-        self.cursor.execute(
+        r = self.writeCursor
+        r.execute(
             'insert into weakrefs (oid_host, oid_ref, ssid)'
             '  values(?, ?, ?)', (oid, oid_ref, self.ssid))
         return oid
 
     def getOrdered(self, oid):
-        r = self.cursor.execute(
+        r = self.readCursor.execute(
             'select '
             '    v_oid, v_stg_kind, v_otype '
             '  from lists_lookup where oid_host=?', (oid,))
         return r.fetchall()
     def setOrdered(self, oid, valueOids):
-        r = self.cursor
+        r = self.writeCursor
         r.execute(
             'delete from mappings '
             '  where oid_host=?', (oid,))
@@ -269,14 +311,14 @@ class SQLStorage(object):
         return oid
 
     def getMapping(self, oid):
-        r = self.cursor.execute(
+        r = self.readCursor.execute(
             'select '
             '    k_oid, k_stg_kind, k_otype, '
             '    v_oid, v_stg_kind, v_otype '
             '  from mappings_lookup where oid_host=?', (oid,))
         return [(e[:3], e[3:]) for e in r.fetchall()]
     def setMapping(self, oid, itemOids):
-        r = self.cursor
+        r = self.writeCursor
         r.execute(
             'delete from mappings '
             '  where oid_host=?', (oid,))
@@ -295,7 +337,8 @@ class SQLStorage(object):
         self.gcRestart()
 
     def gcRestart(self):
-        r = self.cursor
+        r = self.writeCursor
+        if not r: return
         r.execute('''
             delete from oidGraphMembers;''')
         r.execute('''
@@ -303,7 +346,8 @@ class SQLStorage(object):
                 select oid_ref from exports;''')
 
     def gc(self, reap=True):
-        r = self.cursor
+        r = self.writeCursor
+        if not r: return
         delta = self.gcIter(r)
         if not delta and reap:
             return self.gcReap(r)
@@ -338,7 +382,7 @@ class SQLStorage(object):
         return d
 
     def gcReap(self, r):
-        self.commit()
+        self.db.commit()
         count, = r.execute('''select count(oid) from oidGraphMembers;''').fetchone()
         exports, = r.execute('''select count(oid_ref) from exports;''').fetchone()
